// frontend/src/app/admin/projects/models/forms/project-form.model.ts
import { FormControl, FormGroup } from '@angular/forms';

import { PublicationStatus } from '../../../profile/models/dto/profile.dto';
import { ProjectConfidentiality, ProjectRealStatus, ProjectType } from '../dto/project.dto';

export type ProjectForm = FormGroup<{
  id: FormControl<string | null>;
  slug: FormControl<string>;
  projectType: FormControl<ProjectType>;
  realStatus: FormControl<ProjectRealStatus>;
  publicationStatus: FormControl<PublicationStatus>;
  confidentiality: FormControl<ProjectConfidentiality>;
  startDate: FormControl<Date | null>;
  endDate: FormControl<Date | null>;
  ongoing: FormControl<boolean>;
  featured: FormControl<boolean>;
  displayOrder: FormControl<number>;
  demoUrl: FormControl<string>;
  githubUrl: FormControl<string>;
  indexable: FormControl<boolean>;
  skillIds: FormControl<string[]>;
  titleFr: FormControl<string>;
  summaryFr: FormControl<string>;
  descriptionFr: FormControl<string>;
  contextFr: FormControl<string>;
  problemFr: FormControl<string>;
  objectivesFr: FormControl<string>;
  targetUsersFr: FormControl<string>;
  roleFr: FormControl<string>;
  responsibilitiesFr: FormControl<string>;
  solutionFr: FormControl<string>;
  architectureNotesFr: FormControl<string>;
  featuresFr: FormControl<string>;
  challengesFr: FormControl<string>;
  decisionsFr: FormControl<string>;
  resultsFr: FormControl<string>;
  seoTitleFr: FormControl<string>;
  seoDescriptionFr: FormControl<string>;
  titleEn: FormControl<string>;
  summaryEn: FormControl<string>;
  descriptionEn: FormControl<string>;
  contextEn: FormControl<string>;
  problemEn: FormControl<string>;
  objectivesEn: FormControl<string>;
  targetUsersEn: FormControl<string>;
  roleEn: FormControl<string>;
  responsibilitiesEn: FormControl<string>;
  solutionEn: FormControl<string>;
  architectureNotesEn: FormControl<string>;
  featuresEn: FormControl<string>;
  challengesEn: FormControl<string>;
  decisionsEn: FormControl<string>;
  resultsEn: FormControl<string>;
  seoTitleEn: FormControl<string>;
  seoDescriptionEn: FormControl<string>;
}>;